pyproject.toml
[project]
name = "pythonproject-cw"
version = "0.1.0"
description = ""
authors = [
    {name = "Oksana",email = "oksanakysylychak@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-dotenv (>=1.4.2,<2.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)"
]
